requirements_cuda12.txt
diffusers==0.30.0
einops==0.8.0
gradio==4.44.1
httpx==0.27.2
huggingface_hub==0.24.7
jaxtyping==0.2.34
kiui==0.2.10
loguru==0.7.2
lpips==0.1.4
mediapy==1.2.2
numpy<2
omegaconf==2.1.1
open3d==0.18.0
openai==1.51.2
opencv_python==4.10.0.82
opencv_python_headless==4.10.0.82
peft==0.12.0
Pillow==10.4.0
pyacvd==0.2.11
pyfqmr==0.2.1
pygltflib==1.16.2
pyrender==0.1.45
pytorch_lightning==1.5.9
pyvista==0.44.1
scikit-image==0.20.0
torchmetrics==0.11.4
tqdm==4.66.5
transformers==4.41.2
trimesh==3.23.5
tyro==0.8.11
vedo==2024.5.2
xatlas==0.0.9
kaolin==0.16.0
rembg
timm
kornia
-f https://nvidia-kaolin.s3.us-east-2.amazonaws.com/torch-2.1.2_cu121.html
git+https://github.com/sberbank-ai/Real-ESRGAN.git
git+https://github.com/NVlabs/nvdiffrast
git+https://github.com/facebookresearch/pytorch3d.git
